mstflint: Remove compiler flag checks for GCC >= 7
The configure script is checking for the presence of compiler flags that
are only relevant to GCC >= 7. At this time, OL uses a much older
version of GCC so these checks are disabled.

This has been reported to the maintainer.

This is a temporary fix so that an updated package can be provided to
Exadata and will be superseded by a future upstream patch.

Orabug: 27300558

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>