mstflint: Add support for secure boot system
This update adds support for using firmware tools on systems configured
for secure boot. When the system is configured for secure boot, both the
updated userspace tools and a new kernel module, mstflint_access, are
required. See bug 27404202 for more information on mstflint_access.

The updated userspace tools are provided for systems running UEK4 and
UEK5 but the kernel module for secure boot is only available in UEK4 at
this time. The solution for UEK5 is under development.

This version of the package is not based on an official Mellanox release
from the upstream project: https://github.com/Mellanox/mstflint. It is
built off of upstream commit 4d1fabaae6, from the master branch, and
should not be considered fully tested.

Support for sparc64 is not included.

Orabug: 27300558

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>