mstflint: Add Epoch to package versioning
The Epoch tag is added to the spec file to avoid build-time and
install-time conflicts between versions of packages intended for
a particular UEK release or conflicts with RedHat-provided packages.

For build-time, multiple "BuildRequires" tags are used to express
contstraints using the epoch when applicable. This is because the
current tools do not support "BuildConflicts" or the newer Boolean
operators for expressing rich dependencies.

For run-time, the tags: "Conflicts", "Obsoletes", "Provides", and
"Requires" are updated, with the epoch, when needed.

Orabug: 27774601

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>