autotools: Add generated git version
The autogen.sh script produces a file (common/gitversion.h) containing
the git hash of the release for use in the version string of the tools.
For example:

$ mstconfig --version
mstconfig, mstflint 4.8.0, built on May 21 2018, 22:15:41. Git SHA Hash: 4d1faba

In order to produce the file, autogen.sh must be run from within the
mstflint git repo.

We prefer to run autogen.sh (or the equivalent of its contents) and
the configure script as part of the RPM build process but that process
does not take place in the git repo. For this reason, we manually
generate the file so that it has the same git hash as the upstream
release:

$ sh gitversion.sh ./common/gitversion.h

We do this before adding any of our packaging commits.

It would be better to just patch the code to remove everything from the
version string after the mstflint version.

Orabug: 29217219

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>