autotools: Remove calls to macro not supported on OL6
The macro "AX_CHECK_COMPILE_FLAG" is not supported by the version of
autoconf available with OL6. The "autoconf_archives" packages that
contains extra macro definitions is also not available for OL6.

In a previous version of mstflint, the authors indicated that this macro
was being used to:

"Check for new ignore flags needed for GCC >= 7"

so we commented out the checks since we do not use GCC >= 7.

With no better solution at hand, we do the same thing for this new
release.

This is a HACK! The upstream scripts should be fixed to conditionally do
the correct thing for EL6 and EL7 with the base autoconf package
installed.

Orabug: 29217219

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>